+++ CHANGELOG.md
@@ -26,7 +26,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ### Security -->
 <!-- -  -->
 
-## [0.3.0] - 2025-01-19
+## [0.3.1-rd]
+
+### Added
+
+- new option --rubbderduck-folders
+
+### Changed
+
+- Streamlined project setup by extending pyproject.toml and .gitignore, while reducing requirements.txt to the bare minimum that VS Code needs. Thus setup.cfg became superfluous.
+- Some refactoring: handling common cli options now in separate module cli_common.py
+
+
+### Fixed
+
+- 
+
+<!-- -  -->
+<!-- ### Removed -->
+<!-- -  -->
+<!-- ### Security -->
+<!-- -  -->
+
+## [0.3.0]
 
 ### Added